Factor out existing edit script generator. Edit script generator should not assume AST root is in NodeId 0. This should have been in the previous PR. Mea culpa. Rename --edit as --store. Rename default edit script generator. Edit script generators implement a trait. The user can specify an edit script generation algorithm on the command line. Currently only one algorithm is implemented.
